+++ app/message.py
@@ -22,7 +22,7 @@ def get_all_message():
     return Message("getALl", "")
 
 
-def add_game_message(name: str, platform: str, category: str, price: int, score: int, release_year: int) -> Message:
+def add_game_message(name: str, platform: str, category: str, price: float, score: float, release_year: int) -> Message:
     return Message("addGame", {
         'name': name,
         'platform': platform,
